requirements.txt
torch>=2.1.0
torchaudio>=2.1.0
numpy>=2.0.0
scipy>=1.11.0
librosa>=0.10.0
soundfile>=0.12.0
pyrubberband>=0.3.0
pedalboard>=0.9.0
diffusers>=0.25.0
transformers>=4.36.0
laion-clap>=1.1.4
mir_eval>=0.7
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.14.0
